fix(EnvironmentChecker): Improve namespace prefix handling
- Update the logic to correctly identify and handle the namespace prefix based on the current year and month.
- Add a check to ensure the required class is loaded, providing an error message if not.
- Fix messages to correctly indicate the actions taken on file namespace prefixes.

+++ src/Support/EnvironmentChecker.php
@@ -62,26 +62,29 @@ class EnvironmentChecker
     {
         $yearMonth = date('Ym');
 
-        while (202310 <= $yearMonth && !class_exists(\sprintf('MonorepoBuilderPrefix%s\Symfony\Component\Console\Style\SymfonyStyle', $yearMonth))) {
+        while (
+            202310 <= $yearMonth
+            && !class_exists("MonorepoBuilderPrefix$yearMonth\\Symfony\\Component\\Console\\Style\\SymfonyStyle")
+        ) {
             $yearMonth = date('Ym', strtotime('-1 month', strtotime("{$yearMonth}10")));
         }
 
-        echo \PHP_EOL, "The correct namespace prefix is [MonorepoBuilderPrefix$yearMonth].", \PHP_EOL;
+        if (!class_exists("MonorepoBuilderPrefix$yearMonth\\Symfony\\Component\\Console\\Style\\SymfonyStyle")) {
+            echo \PHP_EOL, 'The file [vendor/symplify/monorepo-builder/bootstrap.php] is not loaded.', \PHP_EOL;
+
+            return;
+        }
+
+        echo \PHP_EOL, \sprintf('The namespace prefix should be [%s].', $namespace = "MonorepoBuilderPrefix$yearMonth"), \PHP_EOL;
 
         foreach (array_map('realpath', glob(__DIR__.'/../../{src,tests}{/,/*/,/*/*/,/*/*/*/}*.php', \GLOB_BRACE)) as $file) {
             $contents = file_get_contents($file);
-
-            $replacedContents = preg_replace('/MonorepoBuilderPrefix\d{4}\d{2}/', "MonorepoBuilderPrefix$yearMonth", $contents);
+            $replacedContents = preg_replace('/MonorepoBuilderPrefix\d{4}\d{2}/', $namespace, $contents);
 
             if ($replacedContents !== $contents) {
-                echo "The file's [{$file}] namespace prefix is being fixed...", \PHP_EOL;
+                file_put_contents($file, $replacedContents);
+                echo "The namespace prefix of the file [$file] has been fixed", \PHP_EOL;
             }
-
-            file_put_contents($file, $replacedContents);
         }
-
-        echo 'The all files namespace prefix has been fixed.', \PHP_EOL;
-
-        // exit(0);
     }
 }
